Gene has a method to set the value of the Gene.
This method is used in the Chromosome.java in setAllele() method, to set
the value of a target Gene in a target Chromosome. This solves a major
problem in the whole project! Also get rid of the constructor that takes
an int as parameter to initialize the Gene as that is not needed.