Fix setAllele method in Chromosome.java
Previously setAllele method used `genes.set(i, allele.get(j));` to set
the gene at index i, which actually was a reference of the gene at index
j of allele. So whenever that gene of the allele got mutated or changed,
the gene in the chromosome also got changed. Had to ask SO to solve the
issue caused by this.

'Deep copy' saved the day! Deep copy is brought about by a setValue method
in the Gene class.